rename graph_restriction to graphRestriction to be uniform
