+++ src/biovalidator-core.js
@@ -11,7 +11,7 @@ const request = require("request-promise");
 const AppError = require("./model/application-error");
 const {getFiles, readFile} = require("./utils/file_utils");
 const {isChildTermOf, isValidTerm, isValidTaxonomy} = require("./keywords");
-const GraphRestriction = require("./keywords/graph_restriction");
+const GraphRestriction = require("./keywords/graphRestriction");
 const ValidationError = require("./model/validation-error");
 const logger = require("./winston");
 const {log_error} = require("./utils/logger");
